Rename: 푸터 적용을 위한 폴더 위치 변경
