+++ src/app/(with-footer)/my-center/panel/InfoCard.tsx
