+++ src/app/(with-footer)/my-center/panel/MemberInfoCard.tsx
