+++ src/app/(with-footer)/panel/Footer.tsx
